+++ src/FormControl/FormLabel.tsx
@@ -1,10 +1,12 @@
 import styled, { css } from 'styled-components';
 import type { Theme } from '@sinoui/theme';
 import React from 'react';
+import classNames from 'classnames';
+import mem from '../utils/mem';
 import useFormControlContext from './useFormControlContext';
-import { removeUndefinedProperties } from '../utils/objects';
 import type FormLabelProps from './FormLabelProps';
 import floatingLabelCss from './floatingLabelCss';
+import { FormControlContextData } from './FormControlContext';
 
 /**
  * 获取标签颜色
@@ -44,9 +46,30 @@ const standardCss = css`
   padding-bottom: 4px;
 `;
 
-const FormLabelInner = styled.label.attrs({
-  className: 'sinoui-form-label' as any,
-})<FormLabelProps & { $isAfterRequired: boolean }>`
+type FormLabelInnerPropsType = FormLabelProps & { $isAfterRequired: boolean };
+
+const FormLabelInner = styled.label.attrs(
+  ({
+    className,
+    variant,
+    layout,
+    $isAfterRequired,
+    filled,
+    focused,
+  }: FormLabelInnerPropsType) => ({
+    className: classNames(
+      className,
+      'sinoui-form-label',
+      `sinoui-form-label--variant-${variant ?? 'standard'}`,
+      `sinoui-form-label--layout-${layout ?? 'standard'}`,
+      {
+        'sinoui-form-label--focused': focused,
+        'sinoui-form-label--filled': filled,
+        'sinoui-form-label--required-is-after': $isAfterRequired,
+      },
+    ),
+  }),
+)<FormLabelInnerPropsType>`
   font-size: ${(props) => props.theme.typography.body1.fontSize};
   font-family: ${(props) => props.theme.typography.fontFamily};
   font-weight: ${(props) => props.theme.typography.fontWeightRegular};
@@ -65,37 +88,63 @@ const FormLabelInner = styled.label.attrs({
   ${({ layout }) => layout === 'floating' && floatingLabelCss}
 `;
 
+const mapper: Array<
+  | (keyof FormControlContextData & keyof FormLabelProps)
+  | [keyof FormControlContextData, keyof FormLabelProps]
+> = [
+  'error',
+  'colon',
+  'variant',
+  'filled',
+  'required',
+  'disabled',
+  'focused',
+  'dense',
+  ['labelLayout', 'layout'],
+  ['id', 'htmlFor'],
+];
+
+/**
+ * 从上下文中获取状态
+ *
+ * @param formControlContext 表单控件上下文
+ */
+const getStateFromContext = mem(
+  (formControlContext?: FormControlContextData | null) => {
+    const state: Partial<FormLabelProps> = {};
+    if (!formControlContext) {
+      return state;
+    }
+    mapper.forEach((item) => {
+      const fromKey = Array.isArray(item) ? item[0] : item;
+      const toKey = Array.isArray(item) ? item[1] : item;
+      const value: any = formControlContext[fromKey];
+      if (value != null) {
+        state[toKey] = value;
+      }
+    });
+
+    return state;
+  },
+);
+
 /**
  * 表单标签组件
  *
  * @ReactComponent
  */
 const FormLabel = React.forwardRef<HTMLLabelElement, FormLabelProps>(
-  (props, ref) => {
+  ({ children, ...props }, ref) => {
     const formControlContext = useFormControlContext();
-    const state = formControlContext
-      ? removeUndefinedProperties({
-          error: formControlContext.error,
-          colon: formControlContext.colon,
-          variant: formControlContext.variant,
-          layout: formControlContext.labelLayout,
-          filled: formControlContext.filled,
-          required: formControlContext.required,
-          disabled: formControlContext.disabled,
-          focused: formControlContext.focused,
-          htmlFor: formControlContext.id,
-          dense: formControlContext.dense,
-        })
-      : {};
 
-    return (
-      <FormLabelInner
-        {...state}
-        {...props}
-        ref={ref}
-        $isAfterRequired={formControlContext?.layout !== 'horizontal'}
-      />
-    );
+    const labelProps = {
+      ...getStateFromContext(formControlContext),
+      ...props,
+      ref,
+      $isAfterRequired: formControlContext?.layout !== 'horizontal',
+    };
+
+    return React.createElement(FormLabelInner, labelProps, children);
   },
 );